`PortNamespace`: do not set `dynamic=False` when `valid_type=None`
When setting `valid_type` to anything but `None` the setter will
automatically also set `dynamic` to `True`, because this almost always
what one wants and so saves having to set this attribute explicitly.
Setting a valid type for a namespace but then not allowing any value
whatsoever is non-sensical.

The `valid_type` setter also implemented the reverse, meaning that
setting `valid_type=None` would automatically set `dynamic=False`. This
use case is a lot less common and actually leads to a lot of problems
since the default `valid_type` of a namespace is `None`. When exposing a
namespace that is dynamic, but without a particular `valid_type`, which
is perfectly fine logically - it just accepts any value, the exposed
namespace would lose its dynamicity. This is because the `valid_type`
would be set after the `dynamic`, resetting it to `False` because the
`valid_type` is `None`.